Split script to files
