+++ integrations/aws-v3/scripts/govcloud/__init__.py
@@ -1,0 +1,1 @@
+"""GovCloud setup scripts for the Port AWS v3 integration."""
