+++ integrations/aws-v3/scripts/utils/__init__.py
@@ -1,0 +1,5 @@
+"""General utilities for AWS v3 setup scripts."""
+
+from scripts.utils.logging import logger
+
+__all__ = ["logger"]
